gnn module update
